#Refactored again.  Deleted redundant copies of DraggableCube.js/createDraggableCube.js/createPlayer and createWorld. #Moved all components into compenents folder. #Moved all entities into entities folder.
