+++ entities/DraggableCube.js
